requirements.txt
# pip3 install torch torchaudio --index-url https://download.pytorch.org/whl/cpu
# curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
moviepy
scipy
librosa
soundfile
deepfilternet
openai-whisper
pyannote.audio
pandas
openpyxl
pydub
torch
torchaudio